# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Daily Location Status API
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
  /locations:
    get:
      tags:
        - Locations
      summary: Get all active locations
      responses:
        '200':
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Location'

    post:
      tags:
        - Locations
      summary: Create a new location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Created location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
...
  /reports:
    get:
      tags:
        - Reports
      summary: "Get reports (default: today)"
      parameters:
        - in: query
          name: minDate
          schema:
            type: string
            format: date-time
        - in: query
          name: maxDate
          schema:
            type: string
            format: date-time
        - in: query
          name: date
          schema:
            type: string
            format: date
        - in: query
          name: userId
          schema:
            type: integer
        - in: query
          name: locationId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Report'

    post:
      tags:
        - Reports
      summary: Create a report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, locationId, occuredAt, source]
              properties:
                userId:
                  type: integer
                locationId:
                  type: integer
                isDailyStatusOK:
                  type: boolean
                occuredAt:
                  type: string
                  format: date-time
                source:
                  type: string
                  enum: [ui, bot]
      responses:
        '201':
          description: Created
...
components:
...
  schemas:
    Location:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
...
    Report:
      type: object
      properties:
        userId:
          type: integer
        locationId:
          type: integer
        isDailyStatusOK:
          type: boolean
        occuredAt:
          type: string
          format: date-time